Remove dupe logic
